Add high scores functionality
Fixes #1

I had to go around the Emscripten file system API because I just could not get the thing to work or to tell me why not. As far as I could tell, `syncfs` was failing silently. It was easier and less frustrating to just go around it instead of figuring it out.